Spin key conversion utilities off
